feat: read terminfo database
From the terminfo(5) manual page: Terminfo describes terminals by giving a set
of capabilities which they have, by specifying how to perform screen operations,
and by specifying padding requirements and initialization sequences. This patch
makes use of the tput(1) command to query the 'colors' capability, that the
terminfo(5) manual page describes as the 'maximum number of colors on
the screen'.

Before, this was done by checking the $TERM and $COLORTERM environment
variables, which is not very reliable as terminal emulators may choose to use
non-standard values.